backend/pyproject.toml
[project]
name = "reliable-researcher-backend"
version = "0.1.0"
description = "Backend for Reliable Researcher App"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "langchain>=0.3.14",
    "langgraph>=0.2.64",
    "lancedb>=0.17.0",
    "pypdf>=5.1.0",
    "langchain-google-genai>=2.0.8",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "httpx>=0.28.0"
]
